PHN TECHNOLOGY/Humanoid Prototype/requirements.txt
openai
sounddevice
numpy
scipy
pyserial
pyttsx3
